ADK/L5(customAgent)/requirements.txt
google-generativeai>=0.3.0
requests>=2.31.0
python-dotenv>=1.0.0
PyAudio>=0.2.13
chromadb>=0.4.22
elevenlabs>=0.2.27
pydub>=0.25.1
pydantic>=2.6.1
sentence_transformers>=2.5.1
google-cloud-aiplatform>=1.36.0
google-cloud-storage>=2.10.0
google-cloud-bigquery>=3.11.0
google-cloud-pubsub>=2.18.0
google-cloud-logging>=3.5.0
google-cloud-monitoring>=2.15.0
google-cloud-trace>=1.11.0
google-cloud-error-reporting>=1.5.0
beautifulsoup4>=4.13.4 